buildNextDisplayBoard to displayHelpers.js, no game over checking yet
