+++ client/src/redux/metaActionCreators.js
@@ -20,7 +20,7 @@ export const handleTuTick = id => (
     // increment TU
     dispatch(infoActions.incrementTu());
     // get next display board
-    dispatch(displayActions.buildNextDisplayBoard());
+    dispatch(displayActions.getNextDisplayBoard());
   }
 );
